+++ SourceCode/Data/ListboxItem.cs
@@ -1,4 +1,5 @@
-﻿using System.ComponentModel.DataAnnotations.Schema;
+﻿using Microsoft.AspNetCore.Components;
+using System.ComponentModel.DataAnnotations.Schema;
 
 namespace ModulesRegistry.Data;